requirements.txt
seaborn
numpy
pandas
matplotlib
valohai-utils
statsmodels 
scikit-learn==0.24.2